Update component names to new dot notation
Previously we had things like imager_odgw_assembly. In the new
schema the prefix is now automatically generated as
subsystem.component, so we have decided that the component name
should simply be the suffix of the prefix we had previously been
using once the leading subsystem name has been stripped.

This commit breaks relationships like subscribed values and
commands. Those will be fixed in later commits.